+++ Apps/CesiumViewer/CesiumViewer.js
@@ -39,7 +39,8 @@ define([
         var scene = widget.scene;
         var primitives = scene.getPrimitives();
         var m = new Model('../../../Assets/Model/duck/duck.json');
-//        m.modelMatrix = Matrix4.fromScale(new Cartesian3(900000.0, 900000.0, 900000.0));
+        m.scale = 90000.0;
         primitives.add(m);
+        scene.getPrimitives().setCentralBody(undefined);
     });
 });